Use Market API to check UTXOs
